salesforce-agent-experiment/requirements.txt
langchain>=0.3
langchain-anthropic>=0.3
langchain-core>=0.3
simple-salesforce>=1.12
python-dotenv>=1.0
